behaveTest/requirements.txt
requests
allure-behave
pymongo
urllib3
jsonschema
Naked
pika